document env file support (#230)
This is my suggestion for documenting `.env` support https://github.com/parcel-bundler/parcel/pull/258.

The original suggestion https://github.com/parcel-bundler/website/issues/216 was for this to be listed under the
>advanced section

but most of the advanced section is about authoring plugins/packagers and using the direct api. I thought keeping it alongside the usage explanation, namely post Code Splitting, a better fit.

I may raise an issue in the future to discuss reorganising the documentation as I've found it difficult for a number of reasons in the past.